# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/internal/migrate-public": {
            "post": {
                "security": [
                    {
                        "Authorization": []
                    },
                    {
                        "Access Token": []
                    },
                    {
                        "Entity Key": []
                    },
                    {
                        "User ID": []
                    }
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Database migration",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/score/update": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video Ranking"
                ],
                "summary": "Update Video Score",
                "parameters": [
                    {
                        "type": "string",
                        "description": "VideoID",
                        "name": "x-video-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "EntityID",
                        "name": "x-entity-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "UpdateScoreVideo",
                        "name": "update_score",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateScoreVideo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update video ranking success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/video-global": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video Ranking"
                ],
                "summary": "Get Top Global Videos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Video"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/video-personalized/{entity_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Video Ranking"
                ],
                "summary": "Get Personalized Top Videos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EntityID",
                        "name": "entity_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Video"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.UpdateScoreVideo": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "shares": {
                    "type": "integer"
                },
                "views": {
                    "type": "integer"
                },
                "watch_time": {
                    "type": "integer"
                }
            }
        },
        "model.Video": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "comments": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "length": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "score": {
                    "type": "number"
                },
                "shares": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updater_id": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "watch_time": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "EntityID": {
            "type": "apiKey",
            "name": "x-entity-id",
            "in": "header"
        },
        "VideoID": {
            "type": "apiKey",
            "name": "x-video-id",
            "in": "header"
        }
    }
}